src/Python/requirements.txt
paho-mqtt>=1.6.1,<2.0.0